+++ CHANGES.md
@@ -1,4 +1,13 @@
-# HEKO-1.0.0
+## HEKO-1.0.1
+#### ADDED
+- line between table head and table body on current playlist song list
+- current song details in footer
+#### FIXED
+- play / pause button when listening a song by clicking one on the current playlist
+- scroll bar in current playlist song list
+- red color for closing window button
+
+## HEKO-1.0.0
 #### ADDED
 - sogn queue system
 - shuffle playlist